♻️ refactor: rename opencodego to opencodecodingplan for suffix consistency
Rename internal ID from opencodego → opencodecodingplan to align with
other Coding Plan providers. Display name remains "OpenCode Go".
This allows isCodingPlanProvider() suffix check to work without exceptions.
