+++ packages/model-bank/package.json
@@ -54,7 +54,7 @@
     "./ollama": "./src/aiModels/ollama.ts",
     "./ollamacloud": "./src/aiModels/ollamacloud.ts",
     "./openai": "./src/aiModels/openai.ts",
-    "./opencodeGo": "./src/aiModels/opencodeGo.ts",
+    "./opencodeCodingPlan": "./src/aiModels/opencodeCodingPlan.ts",
     "./opencodeZen": "./src/aiModels/opencodeZen.ts",
     "./openrouter": "./src/aiModels/openrouter.ts",
     "./perplexity": "./src/aiModels/perplexity.ts",
