+++ packages/model-bank/src/aiModels/index.ts
@@ -49,7 +49,7 @@ import { default as nvidia } from './nvidia';
 import { default as ollama } from './ollama';
 import { default as ollamacloud } from './ollamacloud';
 import { default as openai } from './openai';
-import { default as opencodego } from './opencodeGo';
+import { default as opencodecodingplan } from './opencodeCodingPlan';
 import { default as opencodezen } from './opencodeZen';
 import { default as openrouter } from './openrouter';
 import { default as perplexity } from './perplexity';
@@ -151,7 +151,7 @@ export const LOBE_DEFAULT_MODEL_LIST = buildDefaultModelList({
   ollama,
   ollamacloud,
   openai,
-  opencodego,
+  opencodecodingplan,
   opencodezen,
   openrouter,
   perplexity,
@@ -234,7 +234,7 @@ export { default as nvidia } from './nvidia';
 export { default as ollama } from './ollama';
 export { default as ollamacloud } from './ollamacloud';
 export { gptImage1ParamsSchema, default as openai, openaiChatModels } from './openai';
-export { default as opencodego } from './opencodeGo';
+export { default as opencodecodingplan } from './opencodeCodingPlan';
 export { default as opencodezen } from './opencodeZen';
 export { default as openrouter } from './openrouter';
 export { default as perplexity } from './perplexity';
