+++ packages/model-bank/src/modelProviders/index.ts
@@ -50,7 +50,7 @@ import NvidiaProvider from './nvidia';
 import OllamaProvider from './ollama';
 import OllamaCloudProvider from './ollamacloud';
 import OpenAIProvider from './openai';
-import OpenCodeGoProvider from './opencodeGo';
+import OpenCodeCodingPlanProvider from './opencodeCodingPlan';
 import OpenCodeZenProvider from './opencodeZen';
 import OpenRouterProvider from './openrouter';
 import PerplexityProvider from './perplexity';
@@ -217,7 +217,7 @@ export const DEFAULT_MODEL_PROVIDER_LIST = [
   CerebrasProvider,
   ZenMuxProvider,
   OpenCodeZenProvider,
-  OpenCodeGoProvider,
+  OpenCodeCodingPlanProvider,
   StraicoProvider,
   XiaomiMiMoProvider,
   LongCatProvider,
@@ -283,7 +283,7 @@ export { default as NvidiaProviderCard } from './nvidia';
 export { default as OllamaProviderCard } from './ollama';
 export { default as OllamaCloudProviderCard } from './ollamacloud';
 export { default as OpenAIProviderCard } from './openai';
-export { default as OpenCodeGoProviderCard } from './opencodeGo';
+export { default as OpenCodeCodingPlanProviderCard } from './opencodeCodingPlan';
 export { default as OpenCodeZenProviderCard } from './opencodeZen';
 export { default as OpenRouterProviderCard } from './openrouter';
 export { default as PerplexityProviderCard } from './perplexity';
